revert renaming of valuedisplay
