+++ frontend/src/modules/displays/ValueDisplay.tsx
