+++ frontend/src/modules/displays/ValueSelectorDisplay.tsx
