+++ frontend/src/modules/displays/index.ts
@@ -14,5 +14,5 @@ export * from './Carousel';
 export * from './ModeBanner';
 export * from './MultiStepWizard';
 export * from './ToggleValue';
-export * from './ParameterInfo';
-export * from './ParameterValueDisplay';
+export * from './ValueDisplay';
+export * from './ValueSelectorDisplay';
