+++ frontend/src/modules/info/RRInfo.tsx
@@ -15,7 +15,7 @@ import { getCycleMeasurementsRR, getParametersRR } from '../../store/controller/
 import { StoreState } from '../../store/types';
 import { AlarmModal, Knob } from '../controllers';
 import { SettingAdjustProps, ValueModal } from '../controllers/ValueModal';
-import { ValueDisplay, ValueProps } from '../displays/ParameterInfo';
+import { ValueDisplay, ValueProps } from '../displays/ValueDisplay';
 import { BMIN } from './units';
 
 const displaySelector = createStructuredSelector<StoreState, ValueProps>({
